+++ crates/nostr-relay-pool/src/pool/mod.rs
@@ -331,9 +331,9 @@ impl RelayPool {
     }
 
     /// Send multiple events at once to all relays with `WRITE` flag (check [`RelayServiceFlags`] for more details).
-    #[inline]
-    pub async fn batch_event(&self, events: Vec<Event>) -> Result<Output<()>, Error> {
-        self.inner.batch_event(events).await
+    #[deprecated(since = "0.38.0")]
+    pub async fn batch_event(&self, _events: Vec<Event>) -> Result<Output<()>, Error> {
+        unimplemented!()
     }
 
     /// Send event to specific relays
@@ -348,18 +348,18 @@ impl RelayPool {
     }
 
     /// Send multiple events at once to specific relays
-    #[inline]
+    #[deprecated(since = "0.38.0")]
     pub async fn batch_event_to<I, U>(
         &self,
-        urls: I,
-        events: Vec<Event>,
+        _urls: I,
+        _events: Vec<Event>,
     ) -> Result<Output<()>, Error>
     where
         I: IntoIterator<Item = U>,
         U: TryIntoUrl,
         Error: From<<U as TryIntoUrl>::Err>,
     {
-        self.inner.batch_event_to(urls, events).await
+        unimplemented!()
     }
 
     /// Subscribe to filters to all relays with `READ` flag.